add passport registration controller

+++ database/migrations/2019_05_02_140903_create_genres_table.php
@@ -16,8 +16,9 @@ class CreateGenresTable extends Migration
     public function up()
     {
         Schema::create(
-            'genres', function (Blueprint $table) {
-                $table->increments('id');
+            'genres',
+            function (Blueprint $table) {
+                $table->bigIncrements('id');
                 $table->string('title');
                 $table->text('description')->nullable();
