+++ database/migrations/2019_05_02_141204_create_ratings_table.php
@@ -16,8 +16,9 @@ class CreateRatingsTable extends Migration
     public function up()
     {
         Schema::create(
-            'ratings', function (Blueprint $table) {
-                $table->increments('id');
+            'ratings',
+            function (Blueprint $table) {
+                $table->bigIncrements('id');
 
                 $table->timestamps();
             }
